requirements.txt
pytest==5.2.1
selenium==3.14.0
webdriver-manager>=1.8.2
allure-pytest==2.6.0
requests==2.22.0
pyodbc==4.0.27
PyHamcrest==1.9.0
cloudinary==1.18.2
PyMySQL==0.9.3